gdrive-migrator/pyproject.toml
[project]
name = "gdrive-migrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.177.0",
    "google-auth>=2.40.3",
    "pillow>=11.3.0",
    "pillow-heif>=1.1.0",
    "tqdm>=4.67.1",
]
